Changes to schedule layout

+++ pages/schedule/schedule.js
@@ -11,6 +11,7 @@ let selectedDate = null;
 let morningShiftId;
 let afternoonShiftId;
 
+
 export async function initSchedule() {
     createCurrentDate();
     createWeekContainers();
@@ -18,12 +19,6 @@ export async function initSchedule() {
     createCalendar();
     employeeId = await fetchEmployeeId();
   }
-
-function createCurrentDate() {
-  currentDate = new Date();
-  currentDate.setDate(currentDate.getDate() - currentDate.getDay() + 1); // Set to Monday of the current week
-}
-
   async function fetchEmployeeId() {
     const response = await fetch(API_URL + "employee/findbyid", {
       credentials: "include",
@@ -31,6 +26,12 @@ function createCurrentDate() {
     const id = await response.json();
     return id;
   }
+
+function createCurrentDate() {
+  currentDate = new Date();
+  currentDate.setDate(currentDate.getDate() - currentDate.getDay() + 1); // Set to Monday of the current week
+}
+
 
 
   function attachEventListeners() {
@@ -53,6 +54,20 @@ function createCurrentDate() {
     .addEventListener("click", function () {
       cancelShift(afternoonShiftId);
     });
+}
+
+function isDateBeforeToday(date) {
+  const today = new Date();
+  const twelvePM = new Date(today.getFullYear(), today.getMonth(), today.getDate() +1, 12, 0, 0);
+  return (
+    date < twelvePM ||
+    (date.getFullYear() < today.getFullYear()) ||
+    (date.getFullYear() === today.getFullYear() &&
+      date.getMonth() < today.getMonth()) ||
+    (date.getFullYear() === today.getFullYear() &&
+      date.getMonth() === today.getMonth() &&
+      date.getDate() < today.getDate())
+  );
 }
 
 function createWeekContainers() {
@@ -88,15 +103,19 @@ function createCalendar() {
       // Create a new date object to pass to the showModal function
       const clickedDate = new Date(weekStart);
 
+      if (!isDateBeforeToday(clickedDate)) {
       day.addEventListener("click", async () => {
-        const dayElement = document.querySelector(
-          `.day[data-date="${clickedDate.toISOString().substring(0, 10)}"]`
-        );
-        if (dayElement !== null) {
-          await fetchAndDisplayBookings(clickedDate, dayElement, employeeId);
-        }
-        showModal(clickedDate);
-      });
+          const dayElement = document.querySelector(
+            `.day[data-date="${clickedDate.toISOString().substring(0, 10)}"]`
+          );
+          if (dayElement !== null) {
+            await fetchAndDisplayBookings(clickedDate, dayElement, employeeId);
+          }
+          showModal(clickedDate);
+        });
+      } else {
+        day.classList.add("disabled-date"); // This class will be used to style disabled dates
+      }
 
       days.appendChild(day);
       fetchAndDisplayBookings(clickedDate, day, employeeId); // Correctly pass day element here
@@ -110,11 +129,17 @@ async function showModal(date, shiftInfo) {
   const modalDate = document.getElementById("modalDate");
   const closeBtn = document.getElementsByClassName("close")[0];
 
+  getShiftInfo(date, employeeId);
+
   const bookMorningBtn = document.getElementById("bookMorning");
   const bookAfternoonBtn = document.getElementById("bookAfternoon");
-
-  bookMorningBtn.removeAttribute("disabled");
-  bookAfternoonBtn.removeAttribute("disabled");
+  const cancelShiftMorningBtn = document.getElementById("cancelShift1");
+  const cancelShiftAfternoonBtn = document.getElementById("cancelShift2");
+  
+  bookMorningBtn.style.display = "block";
+  cancelShiftMorningBtn.style.display = "none";
+  bookAfternoonBtn.style.display = "block";
+  cancelShiftAfternoonBtn.style.display = "none";
 
   selectedDate = date;
 
@@ -132,19 +157,53 @@ async function showModal(date, shiftInfo) {
 
   modal.style.display = "block";
   modalDate.innerText = date.toDateString();
+  const response = await fetch(
+    API_URL + "booking/findbookingsbydate/" + date.toISOString().split("T")[0],
+    {
+      credentials: "include",
+    }
+  );
+
+  const bookings = await response.json();
+
+  bookings.forEach((booking) => {
+    const shiftStart = new Date(booking.shiftStart);
+    const shiftEnd = new Date(booking.shiftEnd);
+  
+    if (shiftStart.getHours() === 8 && shiftEnd.getHours() === 12) {
+      // Morning shift
+      if (booking.employeeResponse?.employeeId === employeeId) {
+        document.getElementById("bookMorning").style.display = "none";
+        document.getElementById("cancelShift1").style.display = "block";
+      } else {
+        document.getElementById("bookMorning").setAttribute("disabled", "disabled");
+      }
+    } else if (shiftStart.getHours() === 12 && shiftEnd.getHours() === 17) {
+      // Afternoon shift
+      if (booking.employeeResponse?.employeeId === employeeId) {
+        document.getElementById("bookAfternoon").style.display = "none";
+        document.getElementById("cancelShift2").style.display = "block";
+      } else {
+        document.getElementById("bookAfternoon").setAttribute("disabled", "disabled");
+      }
+    }
+  });
+
 
   closeBtn.onclick = function () {
     modal.style.display = "none";
     location.reload();
   };
-
+  
   window.onclick = function (event) {
     if (event.target === modal) {
       modal.style.display = "none";
       location.reload();
     }
   };
+  
 }
+
 
 async function fetchAndDisplayBookings(date, dayElement, employeeId) {
   const response = await fetch(
@@ -178,9 +237,37 @@ async function fetchAndDisplayBookings(date, dayElement, employeeId) {
     }
   });
 
-  dayElement.innerHTML = `<span>${date.getDate()}/${
+  dayElement.innerHTML = `<span>${getDayName(date)} ${date.getDate()}/${
     date.getMonth() + 1
-  }</span><br><span>AM: ${morningShifts} / PM: ${afternoonShifts}</span>`;
+  }</span><br><span>Morning: ${morningShifts} <br>Afternoon: ${afternoonShifts}</span>`;
+  
+  const cancelShiftMorningBtn = document.getElementById("cancelShift1");
+  const cancelShiftAfternoonBtn = document.getElementById("cancelShift2");
+
+  if (currentUserMorningBooked) {
+    document.getElementById("bookMorning").style.display = "none";
+    cancelShiftMorningBtn.style.display = "block";
+  } else {
+    document.getElementById("bookMorning").style.display = "block";
+    cancelShiftMorningBtn.style.display = "none";
+  }
+
+  if (currentUserAfternoonBooked) {
+    document.getElementById("bookAfternoon").style.display = "none";
+    cancelShiftAfternoonBtn.style.display = "block";
+  } else {
+    document.getElementById("bookAfternoon").style.display = "block";
+    cancelShiftAfternoonBtn.style.display = "none";
+  }
+
+  if (currentUserMorningBooked || currentUserAfternoonBooked) {
+    document.getElementById("bookMorning").setAttribute("disabled", "disabled");
+    document.getElementById("bookAfternoon").setAttribute("disabled", "disabled");
+  } else {
+    document.getElementById("bookMorning").removeAttribute("disabled");
+    document.getElementById("bookAfternoon").removeAttribute("disabled");
+  }
+
 
   dayElement.addEventListener("click", () => {
     showModal(date, {
@@ -191,6 +278,11 @@ async function fetchAndDisplayBookings(date, dayElement, employeeId) {
     }),
       getShiftInfo(date, employeeId);
   });
+}
+
+function getDayName(date) {
+  const days = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
+  return days[date.getDay()];
 }
 
 function updateButtonStates(
@@ -223,6 +315,7 @@ function updateButtonStates(
     bookMorningBtn.style.display = "block";
     cancelShiftMorningBtn.style.display = "none";
   }
+
   if (currentUserAfternoonBooked) {
     bookAfternoonBtn.style.display = "none";
     cancelShiftAfternoonBtn.style.display = "block";
@@ -231,6 +324,7 @@ function updateButtonStates(
     cancelShiftAfternoonBtn.style.display = "none";
   }
 }
+
 
 function getWeekNumber(date) {
   const tempDate = new Date(
@@ -352,15 +446,29 @@ async function getShiftInfo(date, employeeId) {
 }
 
 async function cancelShift(shiftId) {
-  console.log("Trying to delete shift with id:  " + shiftId);
+  try {
+    if (!shiftId) {
+      throw new Error("Shift ID is undefined.");
+    }
 
-  const response = await fetch(API_URL + "booking/delete/" + shiftId, {
-    method: "DELETE",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    credentials: "include",
-  });
-  const updatedShiftInfo = await getShiftInfo(selectedDate, employeeId);
-  showModal(selectedDate, updatedShiftInfo);
+    console.log("Trying to delete shift with id: " + shiftId);
+
+    const response = await fetch(API_URL + "booking/delete/" + shiftId, {
+      method: "DELETE",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      credentials: "include",
+    });
+
+    if (!response.ok) {
+      throw new Error("Failed to delete shift.");
+    }
+
+    const updatedShiftInfo = await getShiftInfo(selectedDate, employeeId);
+    showModal(selectedDate, updatedShiftInfo);
+  } catch (error) {
+    console.error(error);
+  }
 }
+
